backend/requirements.txt
asgiref==3.11.1
charset-normalizer==3.3.2
Django==5.1.6
django-cors-headers==4.7.0
djangorestframework==3.15.2
mysqlclient==2.2.7
pillow==12.1.1
reportlab==4.4.9
sqlparse==0.5.5
tzdata==2025.3
